chore: Move some test helpers from TestHelpers_qt to Platform_qt
Slint will reuse Platform_qt at the beginning of the port.
