+++ src/qtcommon/TestHelpers_qt.cpp
@@ -28,57 +28,6 @@ using namespace KDDockWidgets::QtCommon;
 
 namespace KDDockWidgets::Tests {
 
-static QtMessageHandler s_original = nullptr;
-
-static bool shouldBlacklistWarning(const QString &msg, const QString &category)
-{
-    if (category == QLatin1String("qt.qpa.xcb"))
-        return true;
-
-    return msg.contains(QLatin1String("QSocketNotifier: Invalid socket"))
-        || msg.contains(QLatin1String("QWindowsWindow::setGeometry"))
-        || msg.contains(QLatin1String("This plugin does not support"))
-        || msg.contains(QLatin1String("Note that Qt no longer ships fonts"))
-        || msg.contains(QLatin1String("Another dock KDDockWidgets::DockWidget"))
-        || msg.contains(
-            QLatin1String("There's multiple MainWindows, not sure what to do about parenting"))
-        || msg.contains(QLatin1String("Testing::"))
-        || msg.contains(QLatin1String("outside any known screen, using primary screen"))
-        || msg.contains(QLatin1String("Populating font family aliases took"))
-        || msg.contains(QLatin1String("QSGThreadedRenderLoop: expose event received for window"))
-        || msg.contains(QLatin1String("Group.qml:0: ReferenceError: parent is not defined"))
-
-        // Ignore benign warning in Material style when deleting a dock widget. Should be fixed in
-        // Qt.
-        || (msg.contains(QLatin1String("TypeError: Cannot read property"))
-            && msg.contains(QLatin1String("Material")));
-}
-
-static void fatalWarningsMessageHandler(QtMsgType t, const QMessageLogContext &context,
-                                        const QString &msg)
-{
-    if (shouldBlacklistWarning(msg, QLatin1String(context.category)))
-        return;
-    s_original(t, context, msg);
-
-#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
-    // There's a ton of benign warnings in Qt5 QML when app exits, which add maintenance burden.
-    // Please, jump to Qt 6
-    if (Core::Platform::instance()->isQtQuick())
-        return;
-#endif
-
-    if (t == QtWarningMsg) {
-        const std::string expectedWarning = Core::Platform::instance()->m_expectedWarning;
-        if (!expectedWarning.empty() && msg.contains(QString::fromStdString(expectedWarning)))
-            return;
-
-        if (!Platform_qt::isGammaray() && !qEnvironmentVariableIsSet("NO_FATAL")) {
-            std::terminate();
-        }
-    }
-}
-
 static void sleepWithEventLoop(int ms)
 {
     if (!ms)
@@ -94,7 +43,6 @@ static void sleepWithEventLoop(int ms)
 template<typename Func>
 static bool waitFor(Func func, int timeout)
 {
-
     QElapsedTimer timer;
     timer.start();
 
@@ -132,6 +80,12 @@ public:
 
 EventFilter::~EventFilter() = default;
 
+}
+
+bool Platform_qt::tests_wait(int ms) const
+{
+    Tests::sleepWithEventLoop(ms);
+    return true;
 }
 
 bool Platform_qt::tests_waitForWindowActive(Core::Window::Ptr window, int timeout) const
@@ -176,13 +130,6 @@ bool Platform_qt::tests_waitForResize(Core::Controller *c, int timeout) const
     return tests_waitForResize(c->view(), timeout);
 }
 
-bool Platform_qt::tests_waitForEvent(std::shared_ptr<Core::Window> window, QEvent::Type type,
-                                     int timeout) const
-{
-    auto windowqt = static_cast<Window *>(window.get());
-    return tests_waitForEvent(windowqt->qtWindow(), type, timeout);
-}
-
 bool Platform_qt::tests_waitForDeleted(Core::View *view, int timeout) const
 {
     QObject *o = view ? QtCommon::View_qt::asQObject(view) : nullptr;
@@ -201,7 +148,6 @@ bool Platform_qt::tests_waitForDeleted(Core::View *view, int timeout) const
     const bool wasDeleted = !ptr;
     return wasDeleted;
 }
-
 
 bool Platform_qt::tests_waitForDeleted(Core::Controller *o, int timeout) const
 {
@@ -237,65 +183,6 @@ void Platform_qt::tests_doubleClickOn(QPoint globalPos, Core::View *receiver)
     } else {
         // QtWidgets case
         Platform::instance()->sendEvent(receiver, &ev);
-    }
-}
-
-void Platform_qt::installMessageHandler()
-{
-    Tests::s_original = qInstallMessageHandler(Tests::fatalWarningsMessageHandler);
-}
-
-void Platform_qt::uninstallMessageHandler()
-{
-    if (!Tests::s_original)
-        qWarning() << Q_FUNC_INFO
-                   << "No message handler was installed or the fatalWarningsMessageHandler was "
-                      "already uninstalled!";
-    qInstallMessageHandler(Tests::s_original);
-    Tests::s_original = nullptr;
-}
-
-void Platform_qt::tests_initPlatform_impl()
-{
-    qGuiApp->setOrganizationName(QStringLiteral("KDAB"));
-    qGuiApp->setApplicationName(QStringLiteral("dockwidgets-unit-tests"));
-}
-
-void Platform_qt::tests_deinitPlatform_impl()
-{
-    delete qGuiApp;
-}
-
-/*static*/
-QT_BEGIN_NAMESPACE
-extern quintptr Q_CORE_EXPORT qtHookData[];
-QT_END_NAMESPACE
-bool Platform_qt::isGammaray()
-{
-    static bool is = qtHookData[3] != 0;
-    return is;
-}
-
-bool Platform_qt::tests_wait(int ms) const
-{
-    Tests::sleepWithEventLoop(ms);
-    return true;
-}
-
-void Platform_qt::maybeSetOffscreenQPA(int argc, char **argv)
-{
-    bool qpaPassed = false;
-
-    for (int i = 1; i < argc; ++i) {
-        if (qstrcmp(argv[i], "-platform") == 0) {
-            qpaPassed = true;
-            break;
-        }
-    }
-
-    if (!qpaPassed && !qEnvironmentVariableIsSet("KDDW_NO_OFFSCREEN")) {
-        // Use offscreen by default as it's less annoying, doesn't create visible windows
-        qputenv("QT_QPA_PLATFORM", "offscreen");
     }
 }
